# demo-app-api/src/main/resources/OpenAPI-definition.yml
openapi: 3.0.3
info:
  title: Sandbox Objects API
  description: Sandbox Objects API
  version: 1.0.0
servers:
  - url: 'https://localhost:8080/v1'
tags:
  - name: objects
    description: API управления objects
paths:
  /objects:
    post:
      description: Создание нового object
      operationId: createObject
      requestBody:
        description: Запрос на создание нового object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectDataRequest'
      responses:
        201:
          description: Ответ при создании object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectData'
      tags:
        - objects
    get:
      description: Получение списка object
      operationId: getObjects
      parameters:
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
          required: false
          description: Дата начала
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
          required: false
          description: Дата окончания
      responses:
        200:
          description: Отфильтрованный список object
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ObjectData'
      tags:
        - objects
  /objects/{objectId}:
    parameters:
      - in: path
        name: objectId
        description: Идентификатор object
        schema:
          type: string
          format: uuid
        required: true
    get:
      description: Получить object по идентификатору
      operationId: getObjectById
      responses:
        200:
          description: Данные object
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObjectData'
      tags:
        - objects
    put:
      description: Изменение object
      operationId: updateObjectById
      requestBody:
        description: Запрос на изменение object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObjectDataRequest'
      responses:
        200:
          description: OK
      tags:
        - objects
    delete:
      description: Удаление object
      operationId: deleteObjectById
      responses:
        204:
          description: OK
      tags:
        - objects
  /objects/bulk-delete:
    post:
      operationId: bulkDelete
      description: Групповое удаление обхектов
      requestBody:
        description: Запрос для группового удаления
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkDeleteRequest'
      responses:
        204:
          description: OK
      tags:
        - objects

components:
  schemas:
    ObjectData:
      type: object
      description: Сущность "объект"
      properties:
        id:
          description: уникальный идентификатор
          type: string
          format: uuid #for codegen
          readOnly: true
          example: cc0345ee-cb97-11ec-9d64-0242ac120002
        stringField:
          type: string
          example: 'String example'
        intField:
          type: integer
          format: int64
          example: 1
        decimalField:
          type: number
          format: double
          example: 3.14
        jsonField:
          type: string
        createdAt:
          type: string
          format: date-time
          readOnly: true
          example: 2024-07-21T17:32:28Z
        updatedAt:
          type: string
          format: date-time
          readOnly: true
          example: 2024-07-21T17:32:28Z

    ObjectDataRequest:
      type: object
      description: запрос для создания/изменения сущности object
      properties:
        typeId:
          type: integer
          format: int64
          example: 1
        region:
          type: string
          example: Russia
        stringField:
          type: string
          example: 'String example'
        intField:
          type: integer
          format: int64
          example: 1
        decimalField:
          type: number
          format: double
          example: 3.14
        jsonField:
          type: string

    BulkDeleteRequest:
      type: object
      description: Запрос для группового удаления
      properties:
        ids:
          description: Список id на удаление
          type: array
          minItems: 1
          items:
            type: string
            format: uuid

    ExternalObject:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        region:
          type: string